Oups

+++ Common Utilities/ConfigObjects/ItemChance.cs
@@ -3,10 +3,10 @@ namespace Common_Utilities.ConfigObjects
     public class ItemChance
     {
         public string ItemName { get; set; }
-        public float Chance { get; set; }
+        public double Chance { get; set; }
         public string Group { get; set; }
 
-        public void Deconstruct(out string name, out float i, out string groupKey)
+        public void Deconstruct(out string name, out double i, out string groupKey)
         {
             name = ItemName;
             i = Chance;
