+++ Common Utilities/EventHandlers/PlayerHandlers.cs
@@ -92,7 +92,7 @@ namespace Common_Utilities.EventHandlers
             for (int i = 0; i < _plugin.Config.StartingInventories[role].UsedSlots; i++)
             {
                 int r = _plugin.Rng.Next(100);
-                foreach ((string item, float chance, string groupKey) in _plugin.Config.StartingInventories[role][i])
+                foreach ((string item, double chance, string groupKey) in _plugin.Config.StartingInventories[role][i])
                 {
                     if (player != null && !string.IsNullOrEmpty(groupKey) && groupKey != "none" && (!ServerStatic.PermissionsHandler._groups.TryGetValue(groupKey, out var group) || group != player.Group))
                         continue;
